Remove Operation code
Operations are far more complicated than the fancy new async await.
All operations have been replaced by a function in ASCService.
