+++ Sources/ASCKit/AscError.swift
@@ -10,9 +10,11 @@ import Foundation
 public enum AscError: Error {
     case noDataProvided(_ type: String)
     case noUserFound(_ email: String)
+    case noBundleIdFound(_ id: String)
     case noBuildsFound
     case noApiKeysRegistered
     case invalidInput(_ message: String)
     case apiKeyNotFound(_ id: String)
+    case apiKeyActivationFailed(_ key: ApiKey)
     case requestFailed(underlyingErrors: [Error])
 }
